Add CameraHelper

+++ core/src/kiet/nguyentuan/libgdx/demo1/Constants.java
@@ -1,6 +1,6 @@
 package kiet.nguyentuan.libgdx.demo1;
 
 public class Constants {
-    public static final double VIEWPORT_WIDTH=5.0f;
-    public static final double VIEWPORT_HEIGHT=5.0f;
+    public static final float VIEWPORT_WIDTH=5.0f;
+    public static final float VIEWPORT_HEIGHT=5.0f;
 }
